renamed all components
