+++ client/components/MenuBar.js
@@ -4,7 +4,7 @@ import { withStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
-import SwipeableDrawer from './swipeableDrawer';
+import SwipeableDrawer from './SwipeableDrawer';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
